refactor(server): extract PartitionMonitor from EventSource_GENL
Separate partition map and unnamed-device management from the genl
event source so any EventSource can reuse them.

将分区映射与未命名设备管理从 genl 事件源中提取为独立的
PartitionMonitor 类，供多个 EventSource 复用。

Log: 提取 PartitionMonitor 实现 EventSource_GENL 职责分离
PMS: BUG-370779
Influence: EventSource_GENL 不再耦合分区解析与未命名设备管理逻辑，新增 PartitionMonitor 类可被其他 EventSource 复用。
